# k8s/src/appstore-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AppStore API",
    "description": "App catalogue, deployments and teams for the DHBW cloud self-service platform",
    "version": "0.1.0-test.1"
  },
  "paths": {
...
    "/courses/{course}/students": {
      "get": {
        "tags": [
          "Courses"
        ],
        "summary": "List Course Students",
        "description": "List the addresses of the course's ``studierende``, for building teams.\n\nOnly for callers who teach the course (403 ``course_not_taught``\notherwise). 503 if the role provider cannot be reached.",
        "operationId": "listCourseStudents",
        "parameters": [
          {
            "name": "course",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "pattern": "^group:[^#\\s]+$",
              "title": "Course"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Listcoursestudents"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}